Rework add_question to update questions and users
Fixed issue in leaderboard where questions asked was not incremented
after asking a questions. User reducer did not have an action case
to detect when question was asked. Moved this logic from the question
actions to shared, as this action needed to update the user and question
slices of the store.